risolto problema impossibilità di leggere modelli compatibili nella vista di modifica ricambio e, aggiunta la possibilita di vedere e rimuovere le immagine correlate alle varie tabelle

+++ database/migrations/2022_07_18_102804_create_ricambi_ordinati_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->biginteger("testata_id")->unsigned();
             $table->foreign("testata_id")->references("id")->on("testata_ordini")->onDelete('cascade');
             $table->integer("quantita");
-            $table->integer("prezzo_unitario");
+            $table->decimal("prezzo_unitario");
             $table->timestamps();
         });
     }
